requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
Pillow==9.5.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pymongo==4.3.3
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.0.0
pytweening==1.0.7
pywhatkit==5.4
requests==2.30.0
soupsieve==2.4.1
urllib3==2.0.2
Werkzeug==2.3.4
wikipedia==1.4.0
